Rename scripts directory to flytekit_scripts
